UI Fix and Core UI Addition (#18)
* Profile, Setting, and partial RepositoryDetail Core UI

* Padding fix on setting screen
