+++ src/features/profile/screens/ProfileScreen.js
@@ -1,11 +1,14 @@
 // @flow
 
-import React, {Component, Children} from 'react';
-import {View, Text, TouchableOpacity} from 'react-native';
+import React, {Component} from 'react';
+import {View, Text, TouchableOpacity, SafeAreaView} from 'react-native';
 import ParallaxScrollView from 'react-native-parallax-scroll-view';
 import {MaterialIcons} from '@expo/vector-icons';
 
 import type {NavigationScreenProp} from 'react-navigation';
+import DetailsGroup from '../../../general/core-ui/DetailsGroup';
+import DetailsContent from '../../../general/core-ui/DetailsContent';
+import ParallaxButtons from '../../../general/core-ui/ParallaxButtons';
 
 type Props = {
   navigation: NavigationScreenProp<[]>;
@@ -14,172 +17,105 @@ type Props = {
 class ProfileScreen extends Component<Props> {
   render() {
     return (
-      <ParallaxScrollView
-        styles={{height: 10000}}
-        backgroundColor="#272727"
-        contentBackgroundColor="white"
-        parallaxHeaderHeight={300}
-        stickyHeaderHeight={50}
-        contentContainerStyle={styleParallax.contentStyle}
-        renderStickyHeader={() => (
-          <View style={styleParallax.stickyHeader}>
-            <Text style={styleParallax.txtStickyHeader}> sstur </Text>
-          </View>
-        )}
-        renderFixedHeader={() => (
-          <View
-            style={{
-              position: 'absolute',
-              right: 10,
-              top: 10,
-              height: 30,
-              width: 30,
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            <TouchableOpacity
-              onPress={() => this.props.navigation.navigate('Setting')}
+      <SafeAreaView style={{flex: 1}}>
+        <ParallaxScrollView
+          styles={{height: 10000}}
+          backgroundColor="#272727"
+          contentBackgroundColor="white"
+          parallaxHeaderHeight={300}
+          stickyHeaderHeight={65}
+          contentContainerStyle={styleParallax.contentStyle}
+          renderStickyHeader={() => (
+            <View style={styleParallax.stickyHeader}>
+              <Text style={styleParallax.txtStickyHeader}> sstur </Text>
+            </View>
+          )}
+          renderFixedHeader={() => (
+            <View
+              style={{
+                position: 'absolute',
+                right: 10,
+                top: 25,
+                height: 30,
+                width: 30,
+                alignItems: 'center',
+                justifyContent: 'center',
+              }}
             >
-              <MaterialIcons name="settings" size={30} color="white" />
-            </TouchableOpacity>
-          </View>
-        )}
-        renderForeground={() => (
-          <View style={styleParallax.containerForeground}>
-            <View style={styleParallax.containerProfilePicture}>
-              <View style={styleParallax.imgProfilePicture} />
+              <TouchableOpacity
+                onPress={() => this.props.navigation.navigate('Setting')}
+              >
+                <MaterialIcons name="settings" size={30} color="white" />
+              </TouchableOpacity>
             </View>
-            <View style={styleParallax.containerFullName}>
-              <Text style={styleParallax.txtFullName}>Simon Stürmer</Text>
-            </View>
-            <View style={styleParallax.containerUsername}>
-              <Text style={styleParallax.txtUsername}>sstur</Text>
-            </View>
-            <View style={styleParallax.containerButton}>
-              <ParallaxButtons
-                name="Repositories"
-                value={3}
-                onPress={() =>
-                  this.props.navigation.navigate('RepositoryScreen')
-                }
-              />
+          )}
+          renderForeground={() => (
+            <View style={styleParallax.containerForeground}>
+              <View style={styleParallax.containerProfilePicture}>
+                <View style={styleParallax.imgProfilePicture} />
+              </View>
+              <View style={styleParallax.containerFullName}>
+                <Text style={styleParallax.txtFullName}>Simon Stürmer</Text>
+              </View>
+              <View style={styleParallax.containerUsername}>
+                <Text style={styleParallax.txtUsername}>sstur</Text>
+              </View>
+              <View style={styleParallax.containerButton}>
+                <ParallaxButtons
+                  name="Repositories"
+                  value={3}
+                  onPress={() =>
+                    this.props.navigation.navigate('RepositoryScreen')
+                  }
+                />
 
-              <ParallaxButtons
-                name="Stars"
-                value={6}
-                onPress={() =>
-                  this.props.navigation.navigate('RepositoryScreen')
-                }
-              />
-              <ParallaxButtons
-                name="Followers"
-                value={9}
-                onPress={() =>
-                  this.props.navigation.navigate('RepositoryScreen')
-                }
-              />
-              <ParallaxButtons
-                name="Following"
-                value={12}
-                onPress={() =>
-                  this.props.navigation.navigate('RepositoryScreen')
-                }
-              />
+                <ParallaxButtons
+                  name="Stars"
+                  value={6}
+                  onPress={() =>
+                    this.props.navigation.navigate('RepositoryScreen')
+                  }
+                />
+                <ParallaxButtons
+                  name="Followers"
+                  value={9}
+                  onPress={() =>
+                    this.props.navigation.navigate('RepositoryScreen')
+                  }
+                />
+                <ParallaxButtons
+                  name="Following"
+                  value={12}
+                  onPress={() =>
+                    this.props.navigation.navigate('RepositoryScreen')
+                  }
+                />
+              </View>
             </View>
+          )}
+        >
+          <View style={styles.containerProfileDetails}>
+            <DetailsGroup name="Bio">
+              <DetailsContent>Short Biography</DetailsContent>
+            </DetailsGroup>
+            <DetailsGroup name="Website">
+              <DetailsContent>URL</DetailsContent>
+            </DetailsGroup>
+            <DetailsGroup name="Organizations">
+              <DetailsContent>Org 1</DetailsContent>
+              <DetailsContent>Org 2</DetailsContent>
+            </DetailsGroup>
           </View>
-        )}
-      >
-        <View style={styles.containerProfileDetails}>
-          <DetailsGroup name="Bio">
-            <DetailsContent>Short Biography</DetailsContent>
-          </DetailsGroup>
-          <DetailsGroup name="Website">
-            <DetailsContent>URL</DetailsContent>
-          </DetailsGroup>
-          <DetailsGroup name="Organizations">
-            <DetailsContent>Org 1</DetailsContent>
-            <DetailsContent>Org 2</DetailsContent>
-          </DetailsGroup>
-        </View>
-      </ParallaxScrollView>
+        </ParallaxScrollView>
+      </SafeAreaView>
     );
   }
 }
-type ParalaxProps = {
-  name: string;
-  value: number;
-  onPress?: () => boolean;
-};
 
-function ParallaxButtons(props: ParalaxProps) {
-  let {name, value, onPress = null} = props;
-  return (
-    <TouchableOpacity onPress={onPress}>
-      <View style={styleParallax.buttonRepositories}>
-        <Text style={styleParallax.txtButtonValue}> {value} </Text>
-        <Text style={styleParallax.txtButton}> {name} </Text>
-      </View>
-    </TouchableOpacity>
-  );
-}
-type DetailsGroupProps = {
-  name: string;
-  children: ReactEl;
-};
-function DetailsGroup(props: DetailsGroupProps) {
-  let {name, children, ...otherProps} = props;
-  return (
-    <View>
-      <View {...otherProps} style={styles.containerProfileDetailsTitle}>
-        <Text style={styles.txtProfileDetailsTitle}>{name}</Text>
-      </View>
-      <View style={styles.containerProfileDetailsBody}>
-        {Children.map(children, (child, i) => {
-          return (
-            <View
-              key={i}
-              {...otherProps}
-              style={styles.containerProfileDetailsContent}
-            >
-              <Text>{child}</Text>
-            </View>
-          );
-        })}
-      </View>
-    </View>
-  );
-}
-type DetailsContentProps = {
-  children: ReactEl;
-};
-function DetailsContent(props: DetailsContentProps) {
-  let {children, ...otherProps} = props;
-  return <Text {...otherProps}>{children}</Text>;
-}
 const styles = {
   containerProfileDetails: {
     flex: 1,
     flexDirection: 'column',
-  },
-  containerProfileDetailsTitle: {
-    paddingTop: 30,
-    padding: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: 'grey',
-  },
-  containerProfileDetailsBody: {
-    flex: 1,
-    padding: 10,
-    borderBottomWidth: 2,
-    borderBottomColor: 'grey',
-  },
-  containerProfileDetailsContent: {
-    justifyContent: 'center',
-    padding: 10,
-  },
-  txtProfileDetailsTitle: {
-    fontWeight: 'bold',
   },
 };
 const styleParallax = {
@@ -197,7 +133,7 @@ const styleParallax = {
     backgroundColor: '#272727',
     justifyContent: 'center',
     alignItems: 'center',
-    height: 50,
+    height: 80,
   },
   containerProfilePicture: {
     // flex: 1,
@@ -231,30 +167,7 @@ const styleParallax = {
     paddingLeft: 15,
     paddingRight: 15,
   },
-  buttonRepositories: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonStars: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonFollowers: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonFollowing: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
+
   txtStickyHeader: {
     color: 'white',
     fontWeight: 'bold',
@@ -268,15 +181,6 @@ const styleParallax = {
   txtUsername: {
     color: 'white',
     fontSize: 16,
-  },
-  txtButton: {
-    color: 'white',
-    fontSize: 14,
-  },
-  txtButtonValue: {
-    color: 'white',
-    fontWeight: 'bold',
-    fontSize: 24,
   },
 };
 export default ProfileScreen;
